+++ apps/backoffice-v2/src/domains/workflows/fetchers.ts
@@ -226,6 +226,7 @@ export const updateWorkflowDecision = async ({
   body: {
     decision: string | null;
     reason?: string;
+    comment?: string;
   };
   contextUpdateMethod: 'base' | 'director';
 }) => {