+++ apps/backoffice-v2/src/pages/Entity/components/Case/Case.Actions.tsx
@@ -40,8 +40,8 @@ export const Actions: FunctionComponent<IActionsProps> = ({
   } = useCaseActionsLogic({ workflowId: id, fullName });
 
   return (
-    <div className={`sticky top-0 z-50 col-span-2 space-y-2 bg-base-100 px-4 pt-4`}>
-      <div className={`mb-8 flex flex-row justify-between space-x-3.5 pe-[3.35rem]`}>
+    <div className={`col-span-2 space-y-2 bg-base-100 px-4 pt-4`}>
+      <div className={`mb-8 flex flex-row justify-between space-x-3.5`}>
         <AssignDropdown
           assignedUser={assignedUser}
           assignees={assignees}
@@ -53,11 +53,11 @@ export const Actions: FunctionComponent<IActionsProps> = ({
         />
         <CaseOptions />
       </div>
-      <div className={`flex h-20 justify-between`}>
+      <div className={`flex h-20 justify-between gap-4`}>
         <div className={`flex flex-col space-y-3`}>
           <h2
-            className={ctw(`w-[35ch] break-all text-4xl font-semibold leading-9`, {
-              'h-8 w-[24ch] animate-pulse rounded-md bg-gray-200 theme-dark:bg-neutral-focus':
+            className={ctw(`w-full max-w-[35ch] break-all text-4xl font-semibold leading-9`, {
+              'h-8 w-full max-w-[24ch] animate-pulse rounded-md bg-gray-200 theme-dark:bg-neutral-focus':
                 isLoadingCase,
             })}
           >
@@ -68,7 +68,7 @@ export const Actions: FunctionComponent<IActionsProps> = ({
               <span className={`mr-[8px] text-sm leading-6`}>Status</span>
               <Badge
                 variant={tagToBadgeData[tag].variant}
-                className={ctw(`text-sm font-bold`, {
+                className={ctw(`whitespace-nowrap text-sm font-bold`, {
                   'bg-info/20 text-info': tag === StateTag.MANUAL_REVIEW,
                   'bg-violet-500/20 text-violet-500': [
                     StateTag.COLLECTION_FLOW,