+++ packages/workflow-core/src/lib/utils/definition-validator/states-validator.ts
@@ -4,44 +4,51 @@ import { AnyRecord } from '@ballerine/common';
 
 type TTransitionEvent = string;
 
-type TTransitionOption = {
-  target: string;
-  cond?: TDefintionRules;
-};
-type TTransitionOptions = Array<TTransitionOption>;
+type TTransitionOption =
+  | {
+      target: string;
+      cond?: TDefintionRules;
+    }
+  | string;
+type TTransitionOptions = TTransitionOption[];
 export const statesValidator = (
   states: StateMachine<any, any, any>['states'],
   exampleContext?: AnyRecord,
 ) => {
   const stateNames = Object.keys(states);
+
   for (const currentState of stateNames) {
     if (!states[currentState]) {
       throw new Error(`Invalid target state ${currentState} for transition`);
     }
+
     const transitions = states[currentState]!.on;
 
     if (transitions) {
       for (const event of Object.keys(transitions)) {
         const transitionEvent = transitions[event]!;
+
         if (Array.isArray(transitionEvent)) {
           (transitionEvent as unknown as TTransitionOptions).forEach(transitionOption => {
             validateTransitionOnEvent({
               stateNames,
               currentState,
-              targetState: transitionOption.target,
+              transition: transitionOption,
             });
 
-            transitionOption.cond && ruleValidator(transitionOption.cond, exampleContext);
+            if (typeof transitionOption !== 'string' && transitionOption.cond) {
+              ruleValidator(transitionOption.cond, exampleContext);
+            }
           });
+
+          continue;
         }
 
-        if (typeof transitionEvent === 'string') {
-          validateTransitionOnEvent({
-            stateNames,
-            currentState,
-            targetState: transitionEvent as unknown as TTransitionEvent,
-          });
-        }
+        validateTransitionOnEvent({
+          stateNames,
+          currentState,
+          transition: transitionEvent as unknown as TTransitionEvent,
+        });
       }
     }
   }
@@ -50,15 +57,18 @@ export const statesValidator = (
 export const validateTransitionOnEvent = ({
   stateNames,
   currentState,
-  targetState,
+  transition,
 }: {
-  stateNames: Array<string>;
+  stateNames: string[];
   currentState: string;
-  targetState: string;
+  transition: TTransitionOption;
 }) => {
+  const targetState = typeof transition === 'string' ? transition : transition.target;
+
   if (!stateNames.includes(targetState)) {
     throw new Error(`Invalid transition from ${currentState} to ${targetState}`);
   }
+
   if (currentState === targetState) {
     throw new Error(`Recursive transition in state ${targetState}`);
   }