+++ services/workflows-service/src/data-analytics/data-analytics.service.ts
@@ -1,18 +1,24 @@
 import { Injectable } from '@nestjs/common';
 import { PrismaService } from '@/prisma/prisma.service';
 import {
-  InlineRule,
-  TransactionsAgainstDynamicRulesType,
-  TCustomersTransactionTypeOptions,
+  CheckRiskScoreOptions,
   HighTransactionTypePercentage,
-  TPeerGroupTransactionAverageOptions,
-  TDormantAccountOptions,
   HighVelocityHistoricAverageOptions,
+  InlineRule,
+  TCustomersTransactionTypeOptions,
+  TDormantAccountOptions,
+  TPeerGroupTransactionAverageOptions,
+  TransactionsAgainstDynamicRulesType,
+  TMultipleMerchantsOneCounterparty,
+  TExcludedCounterparty,
 } from './types';
 import { AggregateType, TIME_UNITS } from './consts';
-import { Prisma } from '@prisma/client';
+import { AlertSeverity, BusinessReport, BusinessReportType, Prisma } from '@prisma/client';
 import { AppLoggerService } from '@/common/app-logger/app-logger.service';
 import { isEmpty } from 'lodash';
+
+const COUNTERPARTY_ORIGINATOR_JOIN_CLAUSE = Prisma.sql`JOIN "Counterparty" AS "cpOriginator" ON "tr"."counterpartyOriginatorId" = "cpOriginator"."id"`;
+const COUNTERPARTY_BENEFICIARY_JOIN_CLAUSE = Prisma.sql`JOIN "Counterparty" AS "cpBeneficiary" ON "tr"."counterpartyBeneficiaryId" = "cpBeneficiary"."id"`;
 
 @Injectable()
 export class DataAnalyticsService {
@@ -58,6 +64,12 @@ export class DataAnalyticsService {
           ...inlineRule.options,
           projectId,
         });
+
+      case 'evaluateMultipleMerchantsOneCounterparty':
+        return await this[inlineRule.fnName]({
+          ...inlineRule.options,
+          projectId,
+        });
     }
 
     this.logger.error(`No evaluation function found`, {
@@ -65,6 +77,168 @@ export class DataAnalyticsService {
     });
 
     throw new Error(`No evaluation function found for rule name: ${(inlineRule as InlineRule).id}`);
+  }
+
+  private _buildExcludedCounterpartyClause(
+    excludedCounterparty: TExcludedCounterparty = {
+      counterpartyBeneficiaryIds: [],
+      counterpartyOriginatorIds: [],
+    },
+  ) {
+    const excludedCounterpartyClause: {
+      conditions: Prisma.Sql[];
+      join: Prisma.Sql[];
+    } = {
+      conditions: [],
+      join: [],
+    };
+    if (excludedCounterparty) {
+      if (excludedCounterparty.counterpartyBeneficiaryIds.length) {
+        excludedCounterpartyClause.join.push(COUNTERPARTY_BENEFICIARY_JOIN_CLAUSE);
+
+        (excludedCounterparty.counterpartyBeneficiaryIds || []).forEach(id =>
+          excludedCounterpartyClause.conditions.push(
+            Prisma.sql`"cpBeneficiary"."correlationId" NOT LIKE ${id}`,
+          ),
+        );
+      }
+
+      if (excludedCounterparty.counterpartyOriginatorIds.length) {
+        excludedCounterpartyClause.join.push(COUNTERPARTY_ORIGINATOR_JOIN_CLAUSE);
+
+        (excludedCounterparty.counterpartyOriginatorIds || []).forEach(id =>
+          excludedCounterpartyClause.conditions.push(
+            Prisma.sql`"cpOriginator"."correlationId" NOT LIKE ${id}`,
+          ),
+        );
+      }
+    }
+
+    const excludedCounterpartyWhereClause = excludedCounterpartyClause.conditions.length
+      ? Prisma.join(excludedCounterpartyClause.conditions, ' OR ', '(', ')')
+      : Prisma.empty;
+
+    const excludedCounterpartyJoinClause = excludedCounterpartyClause.conditions.length
+      ? Prisma.join(excludedCounterpartyClause.join, '\n ')
+      : Prisma.empty;
+
+    return {
+      excludedCounterpartyClause,
+      excludedCounterpartyWhereClause,
+      excludedCounterpartyJoinClause,
+    };
+  }
+
+  async checkMerchantOngoingAlert(
+    businessReport: BusinessReport,
+    {
+      increaseRiskScorePercentage,
+      increaseRiskScore,
+      maxRiskScoreThreshold,
+    }: CheckRiskScoreOptions,
+    alertSeverity: AlertSeverity,
+  ) {
+    const { report, businessId, projectId } = businessReport;
+
+    if (
+      !(
+        report as {
+          data: {
+            previousReport?: unknown;
+          };
+        }
+      ).data.previousReport
+    ) {
+      return;
+    }
+
+    const {
+      data: {
+        summary: { riskScore: currentRiskScore },
+        previousReport: {
+          summary: { riskScore: previousRiskScore },
+          reportType: previousReportType,
+        },
+      },
+    } = report as {
+      data: {
+        summary: {
+          riskScore: number;
+        };
+        previousReport: {
+          summary: {
+            riskScore: number;
+          };
+          reportType: BusinessReportType;
+        };
+      };
+    };
+
+    if (previousReportType !== BusinessReportType.ONGOING_MERCHANT_REPORT_T1) {
+      this.logger.warn(`Previous report type is not ONGOING_MERCHANT_REPORT_T1`);
+
+      return;
+    }
+
+    if (currentRiskScore < previousRiskScore) {
+      return;
+    }
+
+    if (!(maxRiskScoreThreshold || increaseRiskScore || increaseRiskScorePercentage)) {
+      this.logger.warn(`Rule for ${businessId} ${projectId} missing required options`, {
+        maxRiskScoreThreshold,
+        increaseRiskScore,
+        increaseRiskScorePercentage,
+      });
+
+      return;
+    }
+
+    let ruleResult:
+      | {
+          severity: AlertSeverity;
+          alertReason: string;
+        }
+      | undefined;
+
+    if (maxRiskScoreThreshold && currentRiskScore >= maxRiskScoreThreshold) {
+      ruleResult = {
+        severity: alertSeverity,
+        alertReason: `The risk score has exceeded the threshold of ${maxRiskScoreThreshold}`,
+      };
+    }
+
+    if (increaseRiskScore && currentRiskScore - previousRiskScore >= increaseRiskScore) {
+      ruleResult = {
+        severity: alertSeverity,
+        alertReason: `The risk score has been increased by more than ${increaseRiskScore} from previous monitoring`,
+      };
+    }
+
+    if (
+      increaseRiskScorePercentage &&
+      ((currentRiskScore - previousRiskScore) / previousRiskScore) * 100 >=
+        increaseRiskScorePercentage
+    ) {
+      ruleResult = {
+        severity: alertSeverity,
+        alertReason: `The risk score has been significantly increased from previous monitoring`,
+      };
+    }
+
+    if (!ruleResult) {
+      return;
+    }
+
+    const executionDetails = {
+      businessId: businessId,
+      projectId: projectId,
+      riskScore: currentRiskScore,
+      previousRiskScore,
+      ...ruleResult,
+    };
+
+    return executionDetails;
   }
 
   async evaluateTransactionsAgainstDynamicRules({
@@ -101,7 +275,7 @@ export class DataAnalyticsService {
 
     if (!isEmpty(transactionType)) {
       conditions.push(
-        Prisma.sql`tr."transactionType"::text IN (${Prisma.join([...transactionType], ',')})`,
+        Prisma.sql`"tr"."transactionType"::text IN (${Prisma.join([...transactionType], ',')})`,
       );
     }
 
@@ -175,18 +349,18 @@ export class DataAnalyticsService {
     switch (havingAggregate) {
       case AggregateType.COUNT:
         havingClause = `${AggregateType.COUNT}(id)`;
-        query = Prisma.sql`SELECT ${selectClause}, COUNT(id) AS "transactionCount" FROM "TransactionRecord" tr WHERE ${whereClause} GROUP BY ${groupByClause} HAVING ${Prisma.raw(
+        query = Prisma.sql`SELECT ${selectClause}, COUNT(id) AS "transactionCount" FROM "TransactionRecord" "tr" WHERE ${whereClause} GROUP BY ${groupByClause} HAVING ${Prisma.raw(
           havingClause,
         )} > ${amountThreshold}`;
         break;
       case AggregateType.SUM:
-        havingClause = `${AggregateType.SUM}(tr."transactionBaseAmount")`;
-        query = Prisma.sql`SELECT ${selectClause}, SUM(tr."transactionBaseAmount") AS "totalAmount", COUNT(id) AS "transactionCount" FROM "TransactionRecord" tr WHERE ${whereClause} GROUP BY ${groupByClause} HAVING ${Prisma.raw(
+        havingClause = `${AggregateType.SUM}("tr"."transactionBaseAmount")`;
+        query = Prisma.sql`SELECT ${selectClause}, SUM("tr"."transactionBaseAmount") AS "totalAmount", COUNT(id) AS "transactionCount" FROM "TransactionRecord" "tr" WHERE ${whereClause} GROUP BY ${groupByClause} HAVING ${Prisma.raw(
           havingClause,
         )} > ${amountThreshold}`;
         break;
       default:
-        query = Prisma.sql`SELECT ${selectClause}, COUNT(id) AS "transactionCount" FROM "TransactionRecord" tr WHERE ${whereClause} GROUP BY ${groupByClause}`;
+        query = Prisma.sql`SELECT ${selectClause}, COUNT(id) AS "transactionCount" FROM "TransactionRecord" "tr" WHERE ${whereClause} GROUP BY ${groupByClause}`;
     }
 
     return await this._executeQuery<Array<Record<string, unknown>>>(query);
@@ -239,12 +413,12 @@ export class DataAnalyticsService {
   }) {
     // TODO: get the customer expected amount from the customer's config
     const conditions: Prisma.Sql[] = [
-      Prisma.sql`tr."projectId" = ${projectId}`,
+      Prisma.sql`"tr"."projectId" = ${projectId}`,
       Prisma.sql`jsonb_exists(config, 'customer_expected_amount') AND ((config ->> 'customer_expected_amount')::numeric * ${factor}) != ${customerExpectedAmount}`,
-      Prisma.sql`tr."transactionAmount" > (config ->> 'customer_expected_amount')::numeric`,
+      Prisma.sql`"tr"."transactionAmount" > (config ->> 'customer_expected_amount')::numeric`,
     ];
 
-    const query: Prisma.Sql = Prisma.sql`SELECT tr."businessId" , tr."transactionAmount" FROM "TransactionRecord" as "tr"
+    const query: Prisma.Sql = Prisma.sql`SELECT "tr"."businessId" , "tr"."transactionAmount" FROM "TransactionRecord" as "tr"
     WHERE ${Prisma.join(conditions, ' AND ')} `;
 
     const results = await this.prisma.$queryRaw(query);
@@ -260,21 +434,21 @@ export class DataAnalyticsService {
     const query: Prisma.Sql = Prisma.sql`
   WITH transactions AS (
     SELECT
-      tr."counterpartyBeneficiaryId" AS "counterpartyId",
+      "tr"."counterpartyBeneficiaryId" AS "counterpartyId",
       count(
-        CASE WHEN tr."transactionDate" >= CURRENT_DATE - INTERVAL '${Prisma.raw(
+        CASE WHEN "tr"."transactionDate" >= CURRENT_DATE - INTERVAL '${Prisma.raw(
           `${timeAmount} ${timeUnit}`,
         )}' THEN
-          tr."id"
+          "tr"."id"
         END) AS "totalTransactionWithinSixMonths",
-      count(tr."id") AS "totalTransactionAllTime"
+      count("tr"."id") AS "totalTransactionAllTime"
     FROM
       "TransactionRecord" AS "tr"
-    WHERE 
-      tr."projectId" = ${projectId}
-      AND  tr."counterpartyBeneficiaryId" IS NOT NULL
+    WHERE
+      "tr"."projectId" = ${projectId}
+      AND  "tr"."counterpartyBeneficiaryId" IS NOT NULL
     GROUP BY
-      tr."counterpartyBeneficiaryId"
+      "tr"."counterpartyBeneficiaryId"
   )
   SELECT
     *
@@ -282,7 +456,7 @@ export class DataAnalyticsService {
     transactions
   WHERE
     "totalTransactionAllTime" > 1
-    AND "totalTransactionWithinSixMonths" = 1;  
+    AND "totalTransactionWithinSixMonths" = 1;
   `;
 
     return await this._executeQuery<Array<Record<string, unknown>>>(query);
@@ -307,10 +481,10 @@ export class DataAnalyticsService {
     }
 
     const conditions: Prisma.Sql[] = [
-      Prisma.sql`tr."projectId" = '${projectId}'`,
-      Prisma.sql`tr."businessId" IS NOT NULL`,
+      Prisma.sql`"tr"."projectId" = '${projectId}'`,
+      Prisma.sql`"tr"."businessId" IS NOT NULL`,
       // TODO: should we use equation instead of IN clause?
-      Prisma.sql`tr."transactionType"::text IN (${Prisma.join(transactionType, ',')})`,
+      Prisma.sql`"tr"."transactionType"::text IN (${Prisma.join(transactionType, ',')})`,
       Prisma.sql`"transactionDate" >= CURRENT_DATE - INTERVAL '${Prisma.raw(
         `${timeAmount} ${timeUnit}`,
       )}'`,
@@ -323,7 +497,10 @@ export class DataAnalyticsService {
     // High Velocity - Refund
     const groupBy = {
       clause: Prisma.join(
-        [Prisma.raw(`tr."businessId"`), isPerBrand ? Prisma.raw(`paymentBrandName`) : Prisma.empty],
+        [
+          Prisma.raw(`"tr"."businessId"`),
+          isPerBrand ? Prisma.raw(`paymentBrandName`) : Prisma.empty,
+        ],
         ',',
       ),
     };
@@ -335,7 +512,7 @@ export class DataAnalyticsService {
         havingClause = `${AggregateType.COUNT}(id)`;
         break;
       case AggregateType.SUM:
-        havingClause = `${AggregateType.SUM}(tr."transactionBaseAmount")`;
+        havingClause = `${AggregateType.SUM}("tr"."transactionBaseAmount")`;
         break;
       default:
         throw new Error(`Invalid aggregate type: ${havingAggregate}`);
@@ -366,14 +543,14 @@ export class DataAnalyticsService {
     }
 
     const conditions: Prisma.Sql[] = [
-      Prisma.sql`tr."projectId" = ${projectId}`,
+      Prisma.sql`"tr"."projectId" = ${projectId}`,
       Prisma.sql`"transactionDirection"::text = ${transactionDirection}`,
-      Prisma.sql`tr."paymentMethod"::text ${Prisma.raw(paymentMethod.operator)} ${
+      Prisma.sql`"tr"."paymentMethod"::text ${Prisma.raw(paymentMethod.operator)} ${
         paymentMethod.value
       }`,
       !!timeAmount &&
         !!timeUnit &&
-        Prisma.sql`tr."transactionDate" >= CURRENT_DATE - INTERVAL '${Prisma.raw(
+        Prisma.sql`"tr"."transactionDate" >= CURRENT_DATE - INTERVAL '${Prisma.raw(
           `${timeAmount} ${timeUnit}`,
         )}'`,
       !!customerType && Prisma.sql`b."businessType" = ${customerType}`,
@@ -387,10 +564,10 @@ export class DataAnalyticsService {
         COUNT(*) AS count,
         avg("transactionBaseAmount") AS avg
       FROM
-        "TransactionRecord" tr ${
+        "TransactionRecord" "tr" ${
           customerType
-            ? Prisma.sql`JOIN "Counterparty" AS cp ON tr."counterpartyBeneficiaryId" = cp.id
-               JOIN "Business" AS b ON cp."businessId" = b.id`
+            ? Prisma.sql`JOIN "Counterparty" AS "cp" ON "tr"."counterpartyBeneficiaryId" = "cp".id
+               JOIN "Business" AS b ON "cp"."businessId" = b.id`
             : Prisma.empty
         }
       WHERE
@@ -400,17 +577,17 @@ export class DataAnalyticsService {
       HAVING COUNT(*) > ${minimumCount}
     )
     SELECT
-      tr."counterpartyBeneficiaryId" AS "counterpartyId"
+      "tr"."counterpartyBeneficiaryId" AS "counterpartyId"
     FROM
       "TransactionRecord" tr
-      JOIN "transactionsData" td ON tr."counterpartyBeneficiaryId" = td."counterpartyBeneficiaryId"
+      JOIN "transactionsData" td ON "tr"."counterpartyBeneficiaryId" = td."counterpartyBeneficiaryId"
     WHERE
       "transactionBaseAmount" > ${minimumTransactionAmount}
       AND "transactionBaseAmount" > (
         ${transactionFactor} * avg
       )
     GROUP BY
-      tr."counterpartyBeneficiaryId";
+      "tr"."counterpartyBeneficiaryId";
       `,
     );
   }
@@ -429,23 +606,23 @@ export class DataAnalyticsService {
       throw new Error('projectId is required');
     }
 
-    const historicalTransactionClause = Prisma.sql`tr."transactionDate" >= CURRENT_DATE - INTERVAL '${Prisma.raw(
+    const historicalTransactionClause = Prisma.sql`"tr"."transactionDate" >= CURRENT_DATE - INTERVAL '${Prisma.raw(
       `${activeUserPeriod.timeAmount} ${timeUnit}`,
     )}'`;
 
-    const recentDaysClause = Prisma.sql`tr."transactionDate" >= CURRENT_DATE - INTERVAL '${Prisma.raw(
+    const recentDaysClause = Prisma.sql`"tr"."transactionDate" >= CURRENT_DATE - INTERVAL '${Prisma.raw(
       `${lastDaysPeriod.timeAmount} ${timeUnit}`,
     )}'`;
 
-    const substractPeriodClause = Prisma.sql`tr."transactionDate" >= CURRENT_DATE - INTERVAL '${Prisma.raw(
+    const substractPeriodClause = Prisma.sql`"tr"."transactionDate" >= CURRENT_DATE - INTERVAL '${Prisma.raw(
       `${activeUserPeriod.timeAmount - lastDaysPeriod.timeAmount} ${timeUnit}`,
     )}'`;
 
     const conditions: Prisma.Sql[] = [
-      Prisma.sql`tr."projectId" = ${projectId}`,
-      Prisma.sql`tr."counterpartyBeneficiaryId" IS NOT NULL`,
-      Prisma.sql`tr."transactionDirection"::text = ${transactionDirection}`,
-      Prisma.sql`tr."paymentMethod"::text ${Prisma.raw(paymentMethod.operator)} ${
+      Prisma.sql`"tr"."projectId" = ${projectId}`,
+      Prisma.sql`"tr"."counterpartyBeneficiaryId" IS NOT NULL`,
+      Prisma.sql`"tr"."transactionDirection"::text = ${transactionDirection}`,
+      Prisma.sql`"tr"."paymentMethod"::text ${Prisma.raw(paymentMethod.operator)} ${
         paymentMethod.value
       }`,
       historicalTransactionClause,
@@ -471,43 +648,55 @@ export class DataAnalyticsService {
     ) / 59);
       `,
     );
+  }
 
-    // // TODO: extract count to a variable and resude in select + having section
-    // return await this._executeQuery<Array<{ counterpartyId: string }>>(
-    //   Prisma.sql`
-    // WITH "transactionsData" AS (
-    //   SELECT
-    //     "counterpartyBeneficiaryId",
-    //     COUNT(id) filter (where tr."transactionDate" >= CURRENT_DATE - INTERVAL '${Prisma.raw(
-    //       `${activeUserPeriod.timeAmount} ${activeUserPeriod.timeUnit}`,
-    //     )}')as totalTransactionWithinSixMonths,
-    //     COUNT(id) filter (where tr."transactionDate" >= CURRENT_DATE - INTERVAL '${Prisma.raw(
-    //       `${lastDaysPeriod.timeAmount} ${lastDaysPeriod.timeUnit}`,
-    //     )}')as totalLastDays
-    //   FROM
-    //     "TransactionRecord" tr
-    //   WHERE
-    //     ${Prisma.join(conditions, ' AND ')}
-    //   GROUP BY
-    //     "counterpartyBeneficiaryId"
-    //   HAVING COUNT(
-    //     CASE WHEN tr."transactionDate" >= CURRENT_DATE - INTERVAL '${Prisma.raw(
-    //       `${lastDaysPeriod.timeAmount} ${lastDaysPeriod.timeUnit}`,
-    //     )}' THEN
-    //       tr."id"
-    //     END) > ${minimumCount} -- have to keep the filter of 180 rows
-    // )
-    // SELECT
-    //   tr."counterpartyBeneficiaryId" AS "counterpartyId"
-    // FROM
-    //   "TransactionRecord" tr
-    //   JOIN "transactionsData" td ON tr."counterpartyBeneficiaryId" = td."counterpartyBeneficiaryId"
-    // WHERE
+  async evaluateMultipleMerchantsOneCounterparty({
+    projectId,
+    timeUnit,
+    timeAmount,
+    minimumCount,
+    excludedCounterparty,
+  }: TMultipleMerchantsOneCounterparty) {
+    if (!projectId) {
+      throw new Error('projectId is required');
+    }
 
-    // GROUP BY
-    //   tr."counterpartyBeneficiaryId";
-    //   `,
-    // );
+    const conditions: Prisma.Sql[] = [
+      Prisma.sql`"tr"."projectId" = ${projectId}`,
+      Prisma.sql`"tr"."counterpartyOriginatorId" IS NOT NULL`,
+      Prisma.sql`"cpOriginator"."correlationId" LIKE '%****%'`,
+      !!timeAmount &&
+        !!timeUnit &&
+        Prisma.sql`"tr"."transactionDate" >= CURRENT_DATE - INTERVAL '${Prisma.raw(
+          `${timeAmount} ${timeUnit}`,
+        )}'`,
+    ].filter(Boolean);
+
+    const { excludedCounterpartyWhereClause, excludedCounterpartyClause } =
+      this._buildExcludedCounterpartyClause(excludedCounterparty);
+
+    const join = [COUNTERPARTY_ORIGINATOR_JOIN_CLAUSE, ...excludedCounterpartyClause.join];
+
+    const uniqueJoinMap = new Map(join.map(item => [item.sql, item]));
+    const uniqueJoinClause = [...uniqueJoinMap.values()];
+
+    return await this._executeQuery<Array<{ counterpartyId: string }>>(
+      Prisma.sql`
+      SELECT 
+        "tr"."counterpartyOriginatorId" as "counterpartyId",
+        COUNT(distinct "tr"."counterpartyBeneficiaryId") as "counterpertyInManyBusinessesCount"
+      FROM
+        "TransactionRecord" as "tr" ${Prisma.join(uniqueJoinClause, '\n ')}
+      WHERE
+        ${Prisma.join(
+          [...conditions, excludedCounterpartyWhereClause].filter(cond => cond != Prisma.empty),
+          ' AND ',
+        )}
+      GROUP BY
+        "tr"."counterpartyOriginatorId"
+      HAVING COUNT(distinct "tr"."counterpartyBeneficiaryId") > ${minimumCount};
+      `,
+    );
   }
 
   private async _executeQuery<T = unknown>(query: Prisma.Sql) {