+++ apps/backoffice-v2/CHANGELOG.md
@@ -1,10 +1,20 @@
 # @ballerine/backoffice-v2
 
+## 0.7.9
+
+### Patch Changes
+
+- @ballerine/workflow-browser-sdk@0.6.10
+- @ballerine/workflow-node-sdk@0.6.10
+
 ## 0.7.8
 
 ### Patch Changes
 
 - Updated dependencies
+  - @ballerine/common@0.9.6
+  - @ballerine/workflow-browser-sdk@0.6.9
+  - @ballerine/workflow-node-sdk@0.6.9
   - @ballerine/blocks@0.2.3
 
 ## 0.7.7